requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-apscheduler==0.7.0
APScheduler==3.11.0
requests>=2.31.0
django-cors-headers>=4.3.0

